backend-fastapi/requirements.txt
# Core
fastapi==0.111.0
pydantic==2.11.7
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.30
sqlmodel @ git+https://github.com/tiangolo/sqlmodel@main
asyncpg==0.29.0

# Auth
cryptography==44.0.1
python-jose==3.4.0
bcrypt==4.1.2
#authlib==1.3.1
itsdangerous==2.2.0
httpx==0.27.2

# Server
uvicorn[standard]==0.30.1
gunicorn==23.0.0

# Dev/utility
Faker==25.8.0

user-agents==2.2.0
geoip2==4.8.0
regex>=2024.5.15